Indexed queue
This change implements a new system of tracking the queue, including the ability to delete an entry manually from the current queue.

Other changes are a new master register that keeps track of wether an entry has been entered before and if so, putting it in the Rec (recurring) queue.

NOT TESTED:

Automatic deletion integrated with new queueing system.